WIP: Add chainsaw fill menus
I'm giving each chainsaw and each fuel can an age. In the "fill ..."
menu I'm listing each chainsaw with the amount of fuel it has (and
whether it's equipped) and then under each chainsaw I'm listing each
petrol can with how much fuel it has.

Once a user picks a chainsaw and some fuel then I'll subtract the fuel
"age" from the petrol can and add it to the chainsaw.

Empty petrol cans and full chainsaws won't show up in the menus.

Petrol cans and chainsaws without their fuel "age" in the display will
have it set when the user right clicks.

I might go through and reset chainsaw and petrol can display names every
few minutes or so, ... but I doubt it.