WIP: set the "age names" of chainsaws, gas cans
All the chainsaws and gas cans need to have their ages initialized and
their names set if they aren't already.

I also need to take out the silly {equippedItem, items} pair from
getChainsaws and change it to getItemsByName so I can use it for gas
cans as well. It's easy to get the player's equipped item and check if
an item is equal to it. (Or it is in Java, maybe not in Lua)

The menu seems to almost be working but this code needs to be finished
so that it can smoothly handle named items.