WIP: Fixup items on load
I'm going to yank a lot of this code but I'm keeping a record for
posterity.

It seems that item ages aren't saved but item display names are. When
the game is loaded we trundle through all the chainsaws and petrol cans
and either reload the age from the name, or set the age manually, or set
the age to the default and then update the name accordingly (if
necessary)

I didn't know which events to use where I would have access to the
player so I created a Chainsaw_util.lua script and a bunch of event
handlers to check if a player was available at each event and what types
of parameters the event passes. The results are in event_list.txt.

It turns out that OnLoad, of all the evnets I tried (OnGameBoot,
OnGameStart, OnGameTimeLoaded, OnLoad, OnCreatePlayer) was the last of
the events that fired when I loaded a save game. I'm not sure if this is
the "OnPlayerSaveLoad" or "OnGameLoad" so I might need to tap into
"OnCreatePlayer" as well.