Reference "self" instead of "o"
I think this is stage 6 of the stages of debugging: "how did that ever
work?!"

I'm not sure how "o" was ever in scope. When calling a Lua function with
the colon (:) operator "self" is implied as the first argument. Just
because "o" was the variable name for the object within the "new"
function doesn't mean we can reference "o" outside of that scope. Now
I'm confused.